moved all simple code in FDTests.jl to test snippets. Much cleaner. Still have the more complex code for spherical harmonics to fix.
